Reorganized file structure, fixed several issues from github.
